requirements.txt
streamlit>=1.27.0
openai>=1.3.0
langgraph>=0.0.16
langchain>=0.0.310
langchain-openai>=0.0.5
pydub>=0.25.1
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
moviepy>=1.0.3
numpy>=1.24.0
matplotlib>=3.7.0